Change the dbconnect Redshift user group
When a user connected with Datacampr/DBconnect, a new user was created on Redshift and added to the `readonlyusers` group.
In this PR, we change the users' group to `readonly_dbconnect`. The `readonly_dbconnect` is added to the Redshift's shortquery/fast queue. It should improve morning performance in Belgium/UK for queries issued by a user via DatacampR